pyproject.toml
[project]
name = "AgnoTest"
version = "0.1.0"
description = ""
authors = [
    {name = "Israe KHOUI",email = "israekhoui9@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "streamlit (>=1.52.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "openai (>=2.15.0,<3.0.0)",
    "agno[streamlit] (>=2.3.24,<3.0.0)",
    "mypdf (>=1.0,<2.0)",
    "pypdf (>=6.6.0,<7.0.0)"
]

[tool.poetry]
packages = [{include = "AgnoTest", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
